Address VSA findings: new DatabaseTools.Tests project + split IpcSinks
