+++ tests/Unit/EventLogExpert.DatabaseTools.Tests/GlobalUsings.cs
@@ -1,0 +1,4 @@
+// // Copyright (c) Microsoft Corporation.
+// // Licensed under the MIT License.
+
+global using Xunit;
